+++ lib/install/client/install.go
@@ -30,6 +30,7 @@ import (
 	"github.com/gravitational/gravity/lib/state"
 	"github.com/gravitational/gravity/lib/system/service"
 	"github.com/gravitational/gravity/lib/systemservice"
+	"github.com/gravitational/gravity/lib/utils"
 
 	"github.com/gravitational/trace"
 	log "github.com/sirupsen/logrus"
@@ -86,7 +87,7 @@ func (r *InstallerStrategy) installSelfAsService() error {
 		NoBlock: true,
 		Name:    r.ServicePath,
 	}
-	updateEnviron(&req)
+	req.ServiceSpec.Environment = utils.Getenv(constants.PreflightChecksOffEnvVar)
 	r.WithField("req", fmt.Sprintf("%+v", req)).Info("Install service.")
 	return trace.Wrap(service.Reinstall(req))
 }
@@ -158,18 +159,6 @@ type InstallerStrategy struct {
 	// ConnectTimeout specifies the maximum amount of time to wait for
 	// installer service connection.
 	ConnectTimeout time.Duration
-}
-
-// updateEnviron relays part of the client's environment to service
-func updateEnviron(req *systemservice.NewServiceRequest) {
-	environ := make(map[string]string)
-	for _, env := range []string{constants.PreflightChecksOffEnvVar} {
-		if os.Getenv(env) == "" {
-			continue
-		}
-		environ[env] = os.Getenv(env)
-	}
-	req.ServiceSpec.Environment = environ
 }
 
 var (
